+++ src/app/fs_health.h
@@ -19,14 +19,15 @@ extern "C" {
 typedef struct FSHealthStats {
 		bool storage_partition_present;
 		bool storage_mounted;
-		uint32_t storage_used_bytes;
-		uint32_t storage_total_bytes;
+		uint64_t storage_used_bytes;
+		uint64_t storage_total_bytes;
 } FSHealthStats;
 
 void fs_health_init();
 
-// Called by subsystems that successfully mounted LittleFS.
-void fs_health_set_storage_usage(uint32_t used_bytes, uint32_t total_bytes);
+// Called by subsystems that successfully mounted the storage backend
+// (LittleFS or SD_MMC). uint64_t accommodates SD cards larger than 4 GiB.
+void fs_health_set_storage_usage(uint64_t used_bytes, uint64_t total_bytes);
 
 // Returns cached stats (always succeeds after fs_health_init()).
 void fs_health_get(FSHealthStats* out);
